requirements.txt
Flask==2.0.1
Flask-FlatPages==0.7.3
Markdown==3.4.4
Werkzeug==2.0.3
PyYAML==6.0.1
importlib-metadata==4.13.0 